Extracted button click listener logic from slash command listener.
